st_with_docker/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchvision==0.9.1+cpu
fastai==2.3.1
Pillow==8.2.0
requests==2.25.1
streamlit==0.82.0
toml==0.10.2
urllib3==1.26.4
